Patches deprecated syntax of sha256
This commit patches deprecated syntax of homebrew of `sha256`.

I accidently found out the warning message while updating awless.

Here below is the detail of the warning message I got.

```
Warning: Calling `cellar` in a bottle block is deprecated! Use `brew style --fix` on the formula to update the style or use `sha256` with a `cellar:` argument instead.
```