run "brew style --fix awless.rb"
These warnings were showing up:

Warning: Calling `cellar` in a bottle block is deprecated! Use `brew
style --fix` on the formula to update the style or use `sha256` with a
`cellar:` argument instead.  Please report this issue to the
wallix/awless tap (not Homebrew/brew or Homebrew/core), or even better,
submit a PR to fix it.

Warning: Calling `sha256 "digest" => :tag` in a bottle block is
deprecated! Use `brew style --fix` on the formula to update the style or
use `sha256 tag: "digest"` instead.  Please report this issue to the
wallix/awless tap (not Homebrew/brew or Homebrew/core), or even better,
submit a PR to fix it.